chore: fix coverage
