+++ packages/card/__tests__/shared.base.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Text } from "react-native";
+import TestRenderer from "react-test-renderer";
 import { iterator } from "@times-components/test-utils";
 import Card from "../src/card";
 
@@ -11,7 +12,7 @@ const props = {
   showImage: true
 };
 
-export default (renderMethod, platformTests) => {
+export default renderMethod => {
   // magic to stop the React Native Animated library from dying, as each test kicks off another animation that uses timing
   jest.useFakeTimers();
 
@@ -136,7 +137,106 @@ export default (renderMethod, platformTests) => {
         expect(output).toMatchSnapshot();
       }
     },
-    ...platformTests
+    {
+      name: "card should not re-render when imageRatio prop is changed",
+      test: () => {
+        const testInstance = TestRenderer.create(
+          <Card {...props}>
+            <Text>Do not re-render me</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+
+        testInstance.update(
+          <Card {...props} imageRatio={16 / 9}>
+            <Text>Do not re-render me</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    },
+    {
+      name: "card should re-render when image uri changes",
+      test: () => {
+        const testInstance = TestRenderer.create(
+          <Card {...props}>
+            <Text>Some text</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+
+        testInstance.update(
+          <Card {...props} imageUri="http://foo">
+            <Text>Some text</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    },
+    {
+      name: "card should re-render when low res size changes",
+      test: () => {
+        const testInstance = TestRenderer.create(
+          <Card {...props}>
+            <Text>Some content</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+
+        testInstance.update(
+          <Card {...props} lowResSize={null}>
+            <Text>Some content</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    },
+    {
+      name: "card should re-render when high res size changes",
+      test: () => {
+        const testInstance = TestRenderer.create(
+          <Card {...props}>
+            <Text>Some content</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+
+        testInstance.update(
+          <Card {...props} highResSize={null}>
+            <Text>Some content</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    },
+    {
+      name: "card should re-render when loading state changes",
+      test: () => {
+        const testInstance = TestRenderer.create(
+          <Card {...props} isLoading>
+            <Text>Re-render me</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+
+        testInstance.update(
+          <Card {...props} isLoading={false}>
+            <Text>Re-render me</Text>
+          </Card>
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    }
   ];
 
   iterator(tests);
